Add ability to round values.
Often, a timekeeping system needs time entered as some type of float,
not in hours and minutes. Moreover, those values are often rounded to
timepoints, like every 5, 10, 15, or 30 minutes. This PR adds functionality
to round time so that when the table shows, it provides the float that
would be entered in a timekeeping system. This is an optional argument and
currently available for intervals of 5, 10, 15, and 30 minutes.

Note, with the graph, times are rounded to the specified value; if no
rounding value is specified, for purposes of graphing, they are rounded
to the quarter-hour (i.e., 15 minutes). This is so the y-axis is
intelligible. If `all` is selected with no rounding, the table will
show in time (i.e., HH:MM), but the graph will have time rounded to the
quarter-hour.